re-add bo4e as test-dependency 🙄
