+++ dev_requirements/requirements-tests.in
@@ -1,2 +1,3 @@
 pytest
 pytest-asyncio
+bo4e
