+++ dev_requirements/requirements-tests.txt
@@ -1,18 +1,36 @@
 #
-# This file is autogenerated by pip-compile with Python 3.12
+# This file is autogenerated by pip-compile with Python 3.11
 # by the following command:
 #
 #    pip-compile requirements-tests.in
 #
+annotated-types==0.6.0
+    # via pydantic
+bo4e==202401.0.1
+    # via -r requirements-tests.in
+colorama==0.4.6
+    # via pytest
 iniconfig==2.0.0
     # via pytest
+iso3166==2.1.1
+    # via bo4e
 packaging==23.2
     # via pytest
 pluggy==1.3.0
     # via pytest
+pydantic==2.6.1
+    # via bo4e
+pydantic-core==2.16.2
+    # via pydantic
+pyhumps==3.8.0
+    # via bo4e
 pytest==7.4.4
     # via
-    #   -r dev_requirements/requirements-tests.in
+    #   -r requirements-tests.in
     #   pytest-asyncio
 pytest-asyncio==0.23.4
-    # via -r dev_requirements/requirements-tests.in
+    # via -r requirements-tests.in
+typing-extensions==4.9.0
+    # via
+    #   pydantic
+    #   pydantic-core
